Minor tweaks to fit game data requirements
User stats can include map rank if mapRank is defined

+++ server/src/models/db/map-credit.js
@@ -7,6 +7,6 @@ module.exports = (sequelize, type) => {
 			primaryKey: true,
 			autoIncrement: true
 		},
-		type: type.INTEGER
+		type: type.TINYINT
 	})
 };
